Building 부모 클래스 생성.
- 세계수, 팩토리, 놀이 건물들의 쓰임새가 다르다보니, 공통적인 부분만 묶어서 상속받을 수 있도록 부모 클래스 설계.
- 프로젝트 폴더 정리
