+++ Project-MD/Assets/Images/Building/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a930838c82ee73468598cf94b5ba684
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
