+++ Project-MD/Assets/Images/Building/Materials/WorldTree.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e5a8ebeaaca7c248ae544570d578f43
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
