+++ Project-MD/Assets/Prefabs/Building.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93cd7f4e72bc6a34b82d9456dddd6c7b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
