+++ Project-MD/Assets/Prefabs/Building/Building1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 89beaee306b618b448829dda0b5e1f43
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
