+++ Project-MD/Assets/Prefabs/Building/Building2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 033f6ed6998d448438e5dbc7ea28946b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
