+++ Project-MD/Assets/Prefabs/Building/Building_List_UI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 72886ef5fac04894fae093cb89bb2c84
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
