+++ Project-MD/Assets/Prefabs/Building/WorldTree.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8a6f8d36ef332c845aa43ea28bb38266
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
