--- Project-MD/Assets/Prefabs/Temporary/Building1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 89beaee306b618b448829dda0b5e1f43
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
