--- Project-MD/Assets/Prefabs/Temporary/Building2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 033f6ed6998d448438e5dbc7ea28946b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
