--- Project-MD/Assets/Prefabs/Temporary/Building3.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8a6f8d36ef332c845aa43ea28bb38266
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
