--- Project-MD/Assets/Prefabs/Temporary/Building_List_UI.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 72886ef5fac04894fae093cb89bb2c84
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
