--- Project-MD/Assets/Scripts/Demo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ef8a87fb6f82d384c9442186c7efe4ae
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
