--- Project-MD/Assets/Scripts/Demo/Demo_Building.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b03f5f709e51ab408a2537b3f558259
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
