+++ Project-MD/Assets/Scripts/Map/WorldTree.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+
+public class WorldTree : Building
+{
+    protected override void OnMouseDown()
+    {
+        // UI와 클릭하는 오브젝트가 겹치는 경우.
+        if (EventSystem.current.IsPointerOverGameObject())
+            return;
+
+        if (GameManager.instance.builderManager.get_BuilderMode() == 0) // NO
+        {
+            //GameManager.instance.builderManager.set_BuildingInfo(ID, state, currentPollution);           
+            // 세계수 전용 UI 띄워줌.
+            
+            //
+            return;
+        }
+        else if (GameManager.instance.builderManager.get_BuilderMode() == 1) //Default
+            return;
+        else if (GameManager.instance.builderManager.get_BuilderMode() == 2) // Fix
+        {
+            // Fix 모드인 경우에만 건물 값 세팅.
+            if (isFixed == false)
+                GameManager.instance.builderManager.StartMove(ID, this.gameObject);
+        }
+        else // Relocate
+        {
+            // 건물 값 세팅.
+            GameManager.instance.builderManager.StartMove(ID, this.gameObject);
+        }
+    }
+
+    protected override void OnMouseDrag()
+    {
+        // UI와 클릭하는 오브젝트가 겹치는 경우.
+        if (EventSystem.current.IsPointerOverGameObject())
+            return;
+
+        if (GameManager.instance.builderManager.get_BuilderMode() < 2) // NO, Default
+            return;
+
+        // Fix 모드인 경우에는, 선택한 건물만 움직이도록.
+        if (GameManager.instance.builderManager.get_BuilderMode() == 2) // Fix
+        {
+            if (isFixed == false)
+                GameManager.instance.builderManager.moveItem();
+        }
+        else // Relocate
+            GameManager.instance.builderManager.moveItem();
+    }
+}
